Update `LoadingViewModel`
Return back `BeginInvokeOnMainThread` when changing `MainPage`.

`XmppService_Loaded` method might be called from `DoBind` method,
which in turn is called from `OnAppearing` method. We cannot update
the main page while some `OnAppearing` is still running (well, we can
technically but there will be chaos). Therefore, do not await this
method and do not call it synchronously, even if we are already on
the main thread.