Update `QrCode`
Add an optional parameter to `ScanQrCodeAndHandleResult` method.
It will not change its behavior at the moment because
the method is currently only called from the main experience, when shell
navigation is always used and therefore `UseShellNavigationService`
should be `true`. But the extra parameter will be a hint that
`UseShellNavigationService` might need to be set to `false` (when called
from registration experience).