Update `NavigationService`
Use the page route instead of the page type name when getting
navigation arguments. The reason is that when we push navigation
arguments, we use the route, not the page type name. These two can
differ, e.g. for the chat page on iOS the type name is `ChatPageIos`
and the route is just `ChatPage`.