Update `AppDelegate.cs`
Do not block the UI thread from `OpenUrl` method. `App.OpenUrl` method
posts some work items to the UI thread and `awaits` their completion,
which will deadlock if the UI thread is blocked by the `AppDelegate`'s
`OpenUrl`.

For the time being it will just return `true` but we might need
to add a separate method which will decide if a URL can be handled by
the app without using the UI thread.