Update `ScanQrCodePage`
Now it seems that I have found the true culprit: instead of playing
with the order in which `IsAnalyzing` and `Renderer` are set, we just
need to ensure that `CameraResolutionSelector` is set before we start
actually scanning. The previous code, which tried to "disable"
the scanner and then "enable" it again, didn't really disable / enable
the scanner. It was lucky timing: the old code was executed before the
code which starts scanning.