Update `NavigationService`
Do not store `currentNavigationArgs` as a field but use a computed
property instead. The problem with a field is that it is troublesome
to keep the value of this field in sync with the current navigation
state, especially when we go back multiple pages or perform other
types of non trivial navigation.