Update `ContactListViewModel`
Update `ScanQrCode` method to execute the selection logic
(`this.OnSelected` method) _after_ the QR code page has been dismissed.

The reason why we cannot call `this.OnSelected` method from the QR code
logic using the `Action` parameter is because `this.OnSelected` itself
navigates to a payment page in a manner which creates race condition
between navigating to a payment page and navigating back from the QR
code page.